be/requirements.txt
Django>=4.2.10
djangorestframework>=3.14.0
requests>=2.32.3
pandas>=2.1.0
python-dateutil>=2.8.2
pytz>=2023.3
sqlparse>=0.4.4
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9
